PerlinNoise: double -> float

+++ main/PerlinNoise.hpp
@@ -37,21 +37,21 @@ private:
 
 	std::int32_t p[512];
 
-	static double Fade(double t) noexcept
+	static float Fade(float t) noexcept
 	{
 		return t * t * t * (t * (t * 6 - 15) + 10);
 	}
 
-	static double Lerp(double t, double a, double b) noexcept
+	static float Lerp(float t, float a, float b) noexcept
 	{
 		return a + t * (b - a);
 	}
 
-	static double Grad(std::int32_t hash, double x, double y, double z) noexcept
+	static float Grad(std::int32_t hash, float x, float y, float z) noexcept
 	{
 		const std::int32_t h = hash & 15;
-		const double u = h < 8 ? x : y;
-		const double v = h < 4 ? y : h == 12 || h == 14 ? x : z;
+		const float u = h < 8 ? x : y;
+		const float v = h < 4 ? y : h == 12 || h == 14 ? x : z;
 		return ((h & 1) == 0 ? u : -u) + ((h & 2) == 0 ? v : -v);
 	}
 
@@ -99,17 +99,17 @@ public:
 		}
 	}
 
-	double noise(double x) const
+	float noise(float x) const
 	{
 		return noise(x, 0.0, 0.0);
 	}
 
-	double noise(double x, double y) const
+	float noise(float x, float y) const
 	{
 		return noise(x, y, 0.0);
 	}
 
-	double noise(double x, double y, double z) const
+	float noise(float x, float y, float z) const
 	{
 		const std::int32_t X = static_cast<std::int32_t>(std::floor(x)) & 255;
 		const std::int32_t Y = static_cast<std::int32_t>(std::floor(y)) & 255;
@@ -119,9 +119,9 @@ public:
 		y -= std::floor(y);
 		z -= std::floor(z);
 
-		const double u = Fade(x);
-		const double v = Fade(y);
-		const double w = Fade(z);
+		const float u = Fade(x);
+		const float v = Fade(y);
+		const float w = Fade(z);
 
 		const std::int32_t A = p[X] + Y, AA = p[A] + Z, AB = p[A + 1] + Z;
 		const std::int32_t B = p[X + 1] + Y, BA = p[B] + Z, BB = p[B + 1] + Z;
@@ -136,10 +136,10 @@ public:
 			Grad(p[BB + 1], x - 1, y - 1, z - 1))));
 	}
 
-	double octaveNoise(double x, std::int32_t octaves) const
+	float octaveNoise(float x, std::int32_t octaves) const
 	{
-		double result = 0.0;
-		double amp = 1.0;
+		float result = 0.0;
+		float amp = 1.0;
 
 		for (std::int32_t i = 0; i < octaves; ++i)
 		{
@@ -151,10 +151,10 @@ public:
 		return result;
 	}
 
-	double octaveNoise(double x, double y, std::int32_t octaves) const
+	float octaveNoise(float x, float y, std::int32_t octaves) const
 	{
-		double result = 0.0;
-		double amp = 1.0;
+		float result = 0.0;
+		float amp = 1.0;
 
 		for (std::int32_t i = 0; i < octaves; ++i)
 		{
@@ -167,10 +167,10 @@ public:
 		return result;
 	}
 
-	double octaveNoise(double x, double y, double z, std::int32_t octaves) const
+	float octaveNoise(float x, float y, float z, std::int32_t octaves) const
 	{
-		double result = 0.0;
-		double amp = 1.0;
+		float result = 0.0;
+		float amp = 1.0;
 
 		for (std::int32_t i = 0; i < octaves; ++i)
 		{
@@ -184,32 +184,32 @@ public:
 		return result;
 	}
 
-	double noise0_1(double x) const
+	float noise0_1(float x) const
 	{
 		return noise(x) * 0.5 + 0.5;
 	}
 
-	double noise0_1(double x, double y) const
+	float noise0_1(float x, float y) const
 	{
 		return noise(x, y) * 0.5 + 0.5;
 	}
 
-	double noise0_1(double x, double y, double z) const
+	float noise0_1(float x, float y, float z) const
 	{
 		return noise(x, y, z) * 0.5 + 0.5;
 	}
 
-	double octaveNoise0_1(double x, std::int32_t octaves) const
+	float octaveNoise0_1(float x, std::int32_t octaves) const
 	{
 		return octaveNoise(x, octaves) * 0.5 + 0.5;
 	}
 
-	double octaveNoise0_1(double x, double y, std::int32_t octaves) const
+	float octaveNoise0_1(float x, float y, std::int32_t octaves) const
 	{
 		return octaveNoise(x, y, octaves) * 0.5 + 0.5;
 	}
 
-	double octaveNoise0_1(double x, double y, double z, std::int32_t octaves) const
+	float octaveNoise0_1(float x, float y, float z, std::int32_t octaves) const
 	{
 		return octaveNoise(x, y, z, octaves) * 0.5 + 0.5;
 	}
